# reduce the number of fp.Sync() calls for a WAL write from 3 to 1