--- a/go.mod
+++ b/go.mod
@@ -59,4 +59,3 @@ require (
 	gopkg.in/matryer/try.v1 v1.0.0-20150601225556-312d2599e12e
 	gopkg.in/yaml.v2 v2.2.2
 )
-