@Notargets 2020-03-20
Looks great Hitoshi... I reviewed with “approve” ...

----
- Luke

> On Mar 17, 2020, at 5:34 PM, tomo <notifications@github.com> wrote:
> 
> ﻿
> @umitanuki @Notargets can you review the akimura-san PR?
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or unsubscribe.